Port from javascript to typescript.
