+++ .gitignore
@@ -3,3 +3,5 @@
 /test/_agent.js
 /test/build/
 /test/node_modules/
+/.idea/
+/dist/
